refactor: reorganize model imports and enhance structure
- Split models into domain-specific modules for improved organization and clarity.
- Updated __init__.py to include new model imports and maintain backward compatibility.
- Refactored ActivityLog to use a consistent relationship import style.
- Removed deprecated timezone-aware datetime function from models.py, promoting the use of get_utc_now from base module.
